Undo rename
