+++ packages/mobile/src/components/usdc-balance-pill/USDCBalancePill.tsx
